munch_catering_backend/munch_catering_backend/auth_utils.py
from datetime import timedelta
from jose import jwt
from passlib.context import CryptContext
from munch_catering_backend.settings import settings
from munch_catering_backend.time_utils import utc_now

pwd_context = CryptContext(schemes=["pbkdf2_sha256"], deprecated="auto")

def hash_password(password: str) -> str:
    return pwd_context.hash(password)

def verify_password(plain: str, hashed: str) -> bool:
    return pwd_context.verify(plain, hashed)

def create_token(data: dict) -> str:
    to_encode = data.copy()
    to_encode.update({"exp": utc_now() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
