munch_catering_backend/requirements.txt
fastapi==0.136.0
httpx==0.28.1
motor==3.7.1
passlib==1.7.4
pydantic==2.13.0
email-validator==2.3.0
pymongo==4.16.0
python-dotenv==1.2.2
python-jose==3.5.0
pyjwt[crypto]==2.12.1
python-multipart==0.0.20
